+++ BenchSetup/AssemblyInfo.cs
@@ -1,0 +1,17 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("benchSetup")]
+[assembly: AssemblyDescription("Setup application for Bench")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("Tobias Kiertscher")]
+[assembly: AssemblyProduct("Bench")]
+[assembly: AssemblyCopyright("Copyright © Tobias Kiertscher 2024")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+
+[assembly: ComVisible(false)]
+[assembly: Guid("da29f93f-382f-46fd-93dd-aa0ef6d87bd5")]
+
+[assembly: AssemblyVersion("0.1.0.0")]
+[assembly: AssemblyFileVersion("0.1.0.0")]